+++ lib/utils.dart
@@ -85,47 +85,102 @@ Map<String, Map<String, String>> simulatorsx() {
   return simulators;
 }
 
-/// Creates a list of iOS devices.
+/// Creates a list of available iOS devices.
 /// (really just concerned with simulators for now).
 /// Provides access to their IDs and status'.
 Map getIosDevices() {
-  final deviceInfoRaw = jsonDecode(
-      cmd('xcrun', ['simctl', 'list', 'devices', '--json'], '.', true));
-  final deviceInfo = deviceInfoRaw['devices'];
+  final deviceInfoRaw =
+      cmd('xcrun', ['simctl', 'list', 'devices', '--json'], '.', true);
+  final deviceInfo = jsonDecode(deviceInfoRaw)['devices'];
+  return transformIosDevices(deviceInfo);
+}
 
-  // transform json to a Map of device name by a map of OS's by a list of
+/// Transforms latest information about iOS devices into more convenient
+/// format to index into by device name.
+/// (also useful for testing)
+Map transformIosDevices(deviceInfo) {
+  // transform json to a Map of device name by a map of iOS versions by a list of
   // devices with a map of properties
   // ie, Map<String, Map<String, List<Map<String, String>>>>
+  // In other words, just pop-out the device name for 'easier' access to
+  // the device properties.
   Map deviceInfoTransformed = {};
 
-  deviceInfo.forEach((os, devices) {
+  deviceInfo.forEach((iOSName, devices) {
+    //    print('iOSVersionName=$iOSVersionName');
+    // note: 'isAvailable' does not appear consistently
+    //       so using 'availability' instead
+    isDeviceAvailable(device) => device['availability'] == '(available)';
     for (var device in devices) {
-      // init os map if not already present
+      //      print('device=$device');
+      // init iOS versions map if not already present
       if (deviceInfoTransformed[device['name']] == null) {
         deviceInfoTransformed[device['name']] = {};
       }
 
-      // init os's device array if not already present
-      if (deviceInfoTransformed[device['name']][os] == null) {
-        deviceInfoTransformed[device['name']][os] = [];
+      // init iOS version device array if not already present
+      // note: there can be multiple versions of a device with the same name
+      //       for an iOS version, hence the use of an array.
+      // note: no need to add iOS version if device not available
+      if (deviceInfoTransformed[device['name']][iOSName] == null &&
+          isDeviceAvailable(device)) {
+        deviceInfoTransformed[device['name']][iOSName] = [];
       }
 
-      // add device to os's device array
-      deviceInfoTransformed[device['name']][os].add(device);
+      // add device to iOS version device array
+      // ignoring devices that are not available
+      if (isDeviceAvailable(device))
+        deviceInfoTransformed[device['name']][iOSName].add(device);
     }
   });
   return deviceInfoTransformed;
 }
 
-Map getFirstIosDevice(Map iosDevices, String deviceName) {
-  final oss = iosDevices[deviceName];
-  final osName = oss.keys.first;
-  return iosDevices[deviceName][osName][0];
+// finds the iOS device with the highest available iOS version
+Map getHighestIosDevice(Map iosDevices, String deviceName) {
+  final Map iOSVersions = iosDevices[deviceName];
+
+  // get highest iOS version
+  var iOSVersionName = getHighestIosVersion(iOSVersions);
+
+  final iosVersionDevices = iosDevices[deviceName][iOSVersionName];
+  if (iosVersionDevices.length == 0)
+    throw "Error: no available devices found for \'$deviceName\'";
+  // use the first device found for the iOS version
+  return iosVersionDevices[0];
+}
+
+// returns name of highest iOS version names
+String getHighestIosVersion(Map iOSVersions) {
+  // sort keys in iOS version order
+  final iosVersionNames = iOSVersions.keys.toList();
+//  print('keys=$iosVersionKeys');
+  iosVersionNames.sort((v1, v2) {
+    return v1.compareTo(v2);
+  });
+//  print('keys (sorted)=$iosVersionKeys');
+
+  // get the highest iOS version
+  final iOSVersionName = iosVersionNames.last;
+  return iOSVersionName;
 }
 
 /// Create list of emulators
 List<String> emulators() {
   return cmd('emulator', ['-list-avds'], '.', true).split('\n');
+}
+
+/// Find the android device with the highest available android version
+String getHighestAndroidDevice(String deviceName) {
+  final deviceNameNormalized = deviceName.replaceAll(' ', '_');
+  final devices =
+      emulators().where((name) => name.contains(deviceNameNormalized)).toList();
+  // sort list in android API order
+  devices.sort((v1, v2) {
+    return v1.compareTo(v2);
+  });
+
+  return devices.last;
 }
 
 /// Adds prefix to all files in a directory