+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.7
+- Added check for highest available android emulator for a device
+- Added check for 'adb' in PATH
+
+## 0.1.6
+- Improved handling of iOS simulator info provided by Apple
+
 ## 0.1.5
 
 - Updated parser for iOS simulators to work with all Apple machines